# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API za praćenje kompetencija",
        "description": "Dokumentacija obuhvata rute za upravljanje kompetencijama, verifikacijama i autentifikaciju korisnika.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Prijava korisnika",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešna prijava"
                    },
                    "401": {
                        "description": "Pogrešni podaci"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registracija novog korisnika",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Pera"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Peric"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "pera@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešna registracija"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                }
            }
        },
        "/api/competencies": {
            "get": {
                "tags": [
                    "Competencies"
                ],
                "summary": "Lista kompetencija",
                "operationId": "7c93d80f4e15346f0b90cdde5eaaca4f",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Niste ulogovani"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Competencies"
                ],
                "summary": "Kreiranje nove kompetencije",
                "operationId": "467f5c3b6b8ac15f7f0dde9f70c269c7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "PHP Laravel"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "institution_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "source_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Kompetencija uspešno kreirana"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/competencies/{id}": {
            "get": {
                "tags": [
                    "Competencies"
                ],
                "summary": "Prikaz jedne kompetencije",
                "operationId": "6f8b20052328dd0450ced05def6aefa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji kompetencije"
                    },
                    "404": {
                        "description": "Nije pronađeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Competencies"
                ],
                "summary": "Izmena postojeće kompetencije",
                "operationId": "039106f7c3ba694f5919b6eb858f8953",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Napredni PHP"
                                    },
                                    "level": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno izmenjeno"
                    },
                    "404": {
                        "description": "Nije pronađeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Competencies"
                ],
                "summary": "Brisanje kompetencije",
                "operationId": "3b78929960bab6ffa500371d00c2d0e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Uspešno obrisano"
                    },
                    "404": {
                        "description": "Nije pronađeno"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/all-competencies": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Pregled svih kompetencija u sistemu (Admin)",
                "operationId": "b9464ce3e06d798dca77222815914072",
                "responses": {
                    "200": {
                        "description": "Lista svih kompetencija sa podacima o korisnicima"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}